requirements.txt
PyYAML
requests
twitter
wordnik-py3
